In the proccess of migrating Film to database

+++ WebApi/DTOs/Responses/FilmResponse.cs
@@ -10,10 +10,12 @@ namespace WebApi.DTOs.Responses
         public int FilmId { get; set; }
         public string FilmName { get; set; }
         public string ReleaseDate { get; set; }
-        public int RuntimeInMin { get; set; }
+        public Int16 RuntimeInMin { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
         public string Image { get; set; }
     }
+
+
 }
